Main screen fixes (continued)
Changed the app bar icons, need to fix orientation of stack to be centered, might replace with different widget. Added alternative URLs that reflect more endpoints available (Luis' RPi). Will start doing constrained box or some similar widget for background. Currently having some issues with not responding to close commands as fast, and will try to reflect the globe icon to reflect when app loses connection (HTTP error 503 or anything that is not 200)